// lib/api/types.gen.ts
export type Candle = {
  close?: number | null
  datetime?: number | null
  datetimeISO8601?: string | null
  high?: number | null
  low?: number | null
  open?: number | null
  volume?: number | null
}

export type CandleList = {
  candles?: Array<Candle> | null
  empty?: boolean | null
  previousClose?: number | null
  previousCloseDate?: number | null
  previousCloseDateISO8601?: string | null
  symbol?: string | null
}